Completed ask with futures api and tests. Also reorganised utils into a folder
